[Frontend] Removed box from app and moved layout to MainLayout
