+++ frontend/src/App.tsx
@@ -1,113 +1,24 @@
-import { useContext } from 'react';
+import { Container } from '@mui/material';
 
-import {
-  Box,
-  Button,
-  Card,
-  CardContent,
-  Container,
-  Grid,
-  Stack
-} from '@mui/material';
-
-import { AuthContext, AuthContextProvider } from './components/AuthContext';
-import { CodeEditorWithSyntax } from './components/CodeEditorWithSyntax/CodeEditorWithSyntax';
-import SignUp from './components/SignUp/SignUp';
-import TopAlert from './components/Utils/TopAlert';
-
-import UserService from './services/user.service';
-import CodeHistoryList from './components/CodeHistory/CodeHistoryList';
+import { AuthContextProvider } from './components/AuthContext';
 import { CodeHistoryContextProvider } from './components/contexts/CodeHistoryContext';
-import Chat from './components/Chat/Chat';
-
-const userService = UserService.getInstance();
-
-function Auth() {
-  const { isLoggedIn } = useContext(AuthContext);
-
-  return (
-    <Grid container justifyContent="center" height="100%" spacing={2}>
-      {!isLoggedIn ? (
-        <Grid item>
-          <SignUp />
-        </Grid>
-      ) : (
-        <>
-          <Grid item style={{ textAlign: 'center' }}>
-            <Button
-              variant="contained"
-              color="error"
-              onClick={() => userService.signOut()}
-            >
-              Sign Out
-            </Button>
-          </Grid>
-          <Grid item height="calc(100% - 40px)">
-            <CodeHistoryList />
-          </Grid>
-        </>
-      )}
-    </Grid>
-  );
-}
+import TopAlert from './components/Utils/TopAlert';
+import MainLayout from './components/Layout/mainLayout';
 
 function App() {
   return (
-    <Container
-      style={{ position: 'relative', height: '100vh', paddingBottom: '16px' }}
-    >
+    <>
       <TopAlert />
-      <Box height="100%">
+      <Container
+        style={{ position: 'relative', height: '100vh', paddingBottom: '16px' }}
+      >
         <AuthContextProvider>
           <CodeHistoryContextProvider>
-            <Grid container spacing={2} height="100%" style={{ marginTop: 0 }}>
-              <Grid style={{ height: '100%' }} item xs={12} md={3} lg={3}>
-                <Card style={{ height: '100%' }}>
-                  {/* The padding bottom is 24px, and the top is 16px */}
-                  <CardContent
-                    style={{
-                      position: 'relative',
-                      height: 'calc(100% - 32px)',
-                      paddingBottom: '16px'
-                    }}
-                  >
-                    <Auth />
-                  </CardContent>
-                </Card>
-              </Grid>
-              <Grid item xs={12} md={6} lg={6} overflow={'auto'}>
-                <Card style={{ height: '100%' }}>
-                  {/* The padding bottom is 24px */}
-                  <CardContent
-                    style={{
-                      position: 'relative',
-                      paddingTop: 0,
-                      height: 'calc(100% - 24px)'
-                    }}
-                  >
-                    <CodeEditorWithSyntax />
-                  </CardContent>
-                </Card>
-              </Grid>
-              <Grid style={{ height: '100%' }} item xs={12} md={3} lg={3}>
-                <Card style={{ height: '100%' }}>
-                  {/* The padding bottom is 24px, and the top is 16px */}
-                  <CardContent
-                    style={{
-                      position: 'relative',
-                      height: 'calc(100% - 32px)',
-                      paddingBottom: '16px'
-                    }}
-                  >
-                    <Chat />
-                  </CardContent>
-                </Card>
-              </Grid>
-            </Grid>
+            <MainLayout />
           </CodeHistoryContextProvider>
         </AuthContextProvider>
-      </Box>
-    </Container>
+      </Container>
+    </>
   );
 }
